Format fix.

+++ Questionnaire.Blazor/Models/AnswerModel.cs
@@ -7,7 +7,7 @@
 
         }
 
-        public AnswerModel(int userId, int questionnaireId)
+        public AnswerModel(long userId, int questionnaireId)
         {
             UserId = userId;
             QuestionnaireId = questionnaireId;
@@ -16,7 +16,7 @@
         public int Id { get; set; }
 
         public int QuestionnaireId { get; set; }
-        public int UserId { get; set; }
+        public long UserId { get; set; }
 
         public string Value { get; set; }
     }
